requirements.txt
asgiref==3.7.2
channels==4.0.0
channels-redis==4.2.0
click==8.1.7
daphne==4.0.0
dj-database-url==2.1.0
Django==5.0.1
django-cron==0.6.0
django-crontab==0.7.1
django-redis==5.4.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.122.0
google-cloud-aiplatform==1.45.0
gunicorn==21.2.0
h11==0.14.0
packaging==23.2
psycopg2-binary==2.9.9
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn[standard]==0.27.0
websockets==12.0
whitenoise==6.6.0
